+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
+The format conforms to [Keep a Changelog](http://keepachangelog.com/).
+
+## [Unreleased]
+### Changed
+- Allow `allowShortCircuit` and `allowTernary` for `no-unused-expressions`
 
 ## [1.1.0] - 2016-06-17
 ### Changed
@@ -11,4 +16,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial rule set
 
+[Unreleased]: https://github.com/CoursePark/NestHydrationJS/compare/v1.1.0...HEAD
 [1.1.0]: https://github.com/CoursePark/NestHydrationJS/compare/v1.0.0...v1.1.0